requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.1.0
base58==2.1.1
bitarray==2.9.3
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cmake==3.31.4
cryptography==42.0.8
cytoolz==0.12.3
distro==1.9.0
dotenv==0.9.9
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.0.1
ethereum-abi-utils==0.4.7
ethereum-tester==0.1.0b3
ethereum-utils==0.6.2
exceptiongroup==1.2.2
filelock==3.17.0
flwr==1.12.0
frozenlist==1.5.0
grpcio==1.64.3
h11==0.14.0
hatchling==1.27.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipfshttpclient==0.8.0a2
iterators==0.0.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lit==18.1.8
lru-dict==1.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multiaddr==0.0.9
multidict==6.1.0
netaddr==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.4.0
openai==1.65.4
packaging==24.2
pandas==2.2.3
parsimonious==0.9.0
pathspec==0.12.1
pillow==11.1.0
pluggy==1.5.0
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pylru==1.2.1
pysha3==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.1.0
rpds-py==0.23.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
tomli==2.2.1
tomli_w==1.2.0
toolz==1.0.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tqdm==4.67.1
triton==2.0.0
trove-classifiers==2025.3.19.19
typer==0.12.5
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
varint==1.0.2
web3==7.9.0
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3
